Refactoring: Move the FileDiagnostics from the host package into common.
I am going to reuse the FileDiagnostics in the ProtoBus, so move them into a
common location.

Update the imports.
