+++ proto/host/workspace.proto
@@ -44,34 +44,5 @@ message GetDiagnosticsRequest {
 }
 
 message GetDiagnosticsResponse {
-  repeated FileDiagnostics file_diagnostics = 1;
-}
-
-message FileDiagnostics {
-  string file_path = 1;
-  repeated Diagnostic diagnostics = 2;
-}
-
-message Diagnostic {
-  string message = 1;
-  DiagnosticRange range = 2;
-  DiagnosticSeverity severity = 3;
-  optional string source = 4;
-}
-
-message DiagnosticRange {
-  DiagnosticPosition start = 1;
-  DiagnosticPosition end = 2;
-}
-
-message DiagnosticPosition {
-  int32 line = 1;
-  int32 character = 2;
-}
-
-enum DiagnosticSeverity {
-  DIAGNOSTIC_ERROR = 0;
-  DIAGNOSTIC_WARNING = 1;
-  DIAGNOSTIC_INFORMATION = 2;
-  DIAGNOSTIC_HINT = 3;
+  repeated cline.FileDiagnostics file_diagnostics = 1;
 }
