+++ src/core/mentions/index.ts
@@ -15,7 +15,7 @@ import { openExternal } from "@utils/env"
 import { HostProvider } from "@/hosts/host-provider"
 import { ShowMessageType } from "@/shared/proto/host/window"
 import { diagnosticsToProblemsString } from "@integrations/diagnostics"
-import { DiagnosticSeverity } from "@/shared/proto/index.host"
+import { DiagnosticSeverity } from "@/shared/proto/index.cline"
 
 export async function openMention(mention?: string): Promise<void> {
 	if (!mention) {
